pyproject.toml
[tool.poetry]
name = "example-seed-beetle"
version = "0.0.1"
description = ""
authors = [
  "Luke W. Johnston <lwjohnst@gmail.com>",
  "Kristiane Beicher <kris.beicher@clin.au.dk>",
]
maintainers = [
  "Luke W. Johnston <lwjohnst@gmail.com>",
  "Kristiane Beicher <kris.beicher@clin.au.dk>",
]
readme = "README.md"
repository = "https://github.com/seedcase-project/example-seed-beetle"
license = "MIT"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/seedcase-project/example-seed-beetle/issues"

[tool.poetry.dependencies]
python = "^3.12"
pyjanitor = "^0.30.0"
polars = "^1.22.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
